split configuration and modeling files
